added a new rest api validation

+++ src/main/java/com/really/good/sir/resources/ServiceResource.java
@@ -8,6 +8,7 @@ import com.really.good.sir.dto.ServiceDTO;
 import com.really.good.sir.entity.Role;
 import com.really.good.sir.entity.ServiceEntity;
 import com.really.good.sir.entity.UserSessionEntity;
+import com.really.good.sir.validator.ServiceValidator;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -25,6 +26,7 @@ public class ServiceResource {
     private final ServiceDAO serviceDAO = new ServiceDAO();
     private final ServiceConverter serviceConverter = new ServiceConverter();
     private final UserSessionDAO userSessionDAO = new UserSessionDAO();
+    private final ServiceValidator serviceValidator = new ServiceValidator();
 
     @GET
     public Response getAllServices(@CookieParam("session_id") final String sessionId) {
@@ -58,14 +60,93 @@ public class ServiceResource {
 
     @POST
     public Response createService(ServiceDTO dto, @CookieParam("session_id") final String sessionId) {
+        if (sessionId == null || sessionId.isEmpty()) {
+            final ErrorDTO errorDTO = new ErrorDTO();
+            errorDTO.setMessage("Not authorized");
+            return Response.status(Response.Status.UNAUTHORIZED)
+                    .entity(errorDTO)
+                    .build();
+        }
+        UserSessionEntity session = userSessionDAO.getSessionById(Integer.parseInt(sessionId));
+        if (!Role.ADMIN.toString().equalsIgnoreCase(session.getRole())) {
+            final ErrorDTO errorDTO = new ErrorDTO();
+            errorDTO.setMessage("Forbidden to access resource");
+            return Response.status(Response.Status.FORBIDDEN)
+                    .entity(errorDTO)
+                    .build();
+        }
+
         ServiceEntity entity = serviceConverter.convert(dto);
+        if (!serviceValidator.isNameValid(dto)) {
+            ErrorDTO errorDTO = new ErrorDTO();
+            errorDTO.setMessage("Name has the wrong format");
+            return Response.status(Response.Status.BAD_REQUEST)
+                    .entity(errorDTO)
+                    .build();
+        }
+
+        if (!serviceValidator.isNameUnique(dto)) {
+            ErrorDTO errorDTO = new ErrorDTO();
+            errorDTO.setMessage("Name already exists");
+            return Response.status(Response.Status.BAD_REQUEST)
+                    .entity(errorDTO)
+                    .build();
+        }
+
+        if (!serviceValidator.isPriceValid(dto)) {
+            ErrorDTO errorDTO = new ErrorDTO();
+            errorDTO.setMessage("Price has to be over 0");
+            return Response.status(Response.Status.BAD_REQUEST)
+                    .entity(errorDTO)
+                    .build();
+        }
         ServiceEntity created = serviceDAO.createService(entity);
         return Response.ok(serviceConverter.convert(created)).build();
     }
 
     @PUT
     public Response updateService(ServiceDTO dto, @CookieParam("session_id") final String sessionId) {
+        if (sessionId == null || sessionId.isEmpty()) {
+            final ErrorDTO errorDTO = new ErrorDTO();
+            errorDTO.setMessage("Not authorized");
+            return Response.status(Response.Status.UNAUTHORIZED)
+                    .entity(errorDTO)
+                    .build();
+        }
+        UserSessionEntity session = userSessionDAO.getSessionById(Integer.parseInt(sessionId));
+        if (!Role.ADMIN.toString().equalsIgnoreCase(session.getRole())) {
+            final ErrorDTO errorDTO = new ErrorDTO();
+            errorDTO.setMessage("Forbidden to access resource");
+            return Response.status(Response.Status.FORBIDDEN)
+                    .entity(errorDTO)
+                    .build();
+        }
+
         ServiceEntity entity = serviceConverter.convert(dto);
+        if (!serviceValidator.isNameValid(dto)) {
+            ErrorDTO errorDTO = new ErrorDTO();
+            errorDTO.setMessage("Name has the wrong format");
+            return Response.status(Response.Status.BAD_REQUEST)
+                    .entity(errorDTO)
+                    .build();
+        }
+
+        if (!serviceValidator.isNameUnique(dto)) {
+            ErrorDTO errorDTO = new ErrorDTO();
+            errorDTO.setMessage("Name already exists");
+            return Response.status(Response.Status.BAD_REQUEST)
+                    .entity(errorDTO)
+                    .build();
+        }
+
+        if (!serviceValidator.isPriceValid(dto)) {
+            ErrorDTO errorDTO = new ErrorDTO();
+            errorDTO.setMessage("Price has to be over 0");
+            return Response.status(Response.Status.BAD_REQUEST)
+                    .entity(errorDTO)
+                    .build();
+        }
+
         boolean updated = serviceDAO.updateService(entity);
         LOGGER.info("Service updated: {}", updated);
         return Response.ok(serviceConverter.convert(entity)).build();
@@ -74,6 +155,21 @@ public class ServiceResource {
     @DELETE
     @Path("/{id}")
     public Response deleteService(@PathParam("id") int id, @CookieParam("session_id") final String sessionId) {
+        if (sessionId == null || sessionId.isEmpty()) {
+            final ErrorDTO errorDTO = new ErrorDTO();
+            errorDTO.setMessage("Not authorized");
+            return Response.status(Response.Status.UNAUTHORIZED)
+                    .entity(errorDTO)
+                    .build();
+        }
+        UserSessionEntity session = userSessionDAO.getSessionById(Integer.parseInt(sessionId));
+        if (!Role.ADMIN.toString().equalsIgnoreCase(session.getRole())) {
+            final ErrorDTO errorDTO = new ErrorDTO();
+            errorDTO.setMessage("Forbidden to access resource");
+            return Response.status(Response.Status.FORBIDDEN)
+                    .entity(errorDTO)
+                    .build();
+        }
         boolean deleted = serviceDAO.deleteService(id);
         LOGGER.info("Service deleted: {}", deleted);
         return Response.noContent().build();
